+++ sdk/keyvault/azure-security-keyvault-keys/src/cryptography/decrypt_result.cpp
@@ -4,7 +4,7 @@
 #include <azure/core/internal/json/json.hpp>
 
 #include "../private/key_constants.hpp"
-#include "azure/keyvault/keys/cryptography/cryptography_serializers.hpp"
+#include "../private/cryptography_serializers.hpp"
 #include "azure/keyvault/keys/cryptography/decrypt_result.hpp"
 
 #include <azure/keyvault/common/internal/base64url.hpp>
