+++ sdk/keyvault/azure-security-keyvault-keys/src/cryptography/encrypt_result.cpp
@@ -3,8 +3,8 @@
 
 #include <azure/core/internal/json/json.hpp>
 
+#include "../private/cryptography_serializers.hpp"
 #include "../private/key_constants.hpp"
-#include "azure/keyvault/keys/cryptography/cryptography_serializers.hpp"
 #include "azure/keyvault/keys/cryptography/encrypt_result.hpp"
 
 #include <azure/keyvault/common/internal/base64url.hpp>
