+++ sdk/keyvault/azure-security-keyvault-keys/src/cryptography/key_sign_parameters.cpp
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include "azure/keyvault/keys/cryptography/key_sign_parameters.hpp"
+#include "../private/key_sign_parameters.hpp"
 #include "../private/key_constants.hpp"
-#include "azure/keyvault/keys/cryptography/cryptography_serializers.hpp"
+#include "../private/cryptography_serializers.hpp"
 
 #include <azure/keyvault/common/internal/base64url.hpp>
