+++ sdk/keyvault/azure-security-keyvault-keys/src/private/key_sign_parameters.hpp
